Wire #2091's vtable-partition licensing boundary into CI
#2091 added a fail-closed boundary -- validated_vtable_partition_symbols,
manifest_vtable_partition_rows, and audit_tu_object's validated_vtable_policies
parameter, none of which exist on main before it -- that refuses a
manifest-declared symbol carved out of a vtable's public range unless
partition_vtable_rebiases independently validated it against content-bound
baseline ELF metadata.

Six tests describe that boundary. Five of them never execute:

  test_tu_production.py  1 test   module IS in tool-tests.yml   -> runs
  test_objisolate.py     1 test   module NOT in the list        -> never runs
  test_tubuild.py        4 tests  module runs, finds 0 tests    -> never run

test_tubuild.py has bare `def test_` functions and no unittest.TestCase, so
`python -m unittest tools.test_tubuild` prints "Ran 0 tests ... OK" and exits 0.

That matters now rather than eventually. Measured on #2091's merge tree, the
boundary licenses nothing at all -- 0 of 97 manifest entries, 0
partition_symbols rows tree-wide -- and #2096 introduces the first two rows
that have ever existed (FlyGuy_VT7, FlyGuy_VT14). It goes from unexercised to
load-bearing in one PR, with a decorative safety net.

tools/test_tubuild_vtable_partitions.py -- new, four tests, 4 of 4 run in CI
and none skips. Mock- and dict-driven: no compiler, no ROM, no config/, no
build/. Three of the four are #2091's own toolchain-free tests, moved verbatim
out of test_tubuild.py -- bodies unchanged, `def f():` to `def f(self):` plus
one indent level, plain asserts kept -- so the move reviews as a move.

The fourth is new and pins the invariant that is easiest to re-break:
licensing compares sectionIndex, never section NAMES. A linked ELF names a
vtable's output section after its overlay (OV036/OV047/OV070, measured on the
baseline link) while a manifest section is one of .rodata/.init/.ctor/.data/
.bss. Comparing them refuses every real input, which is what made the whole
path unreachable, and 0b0e319 removed exactly that comparison. Mutation-checked
in both directions:

  drop the sectionIndex invariant  -> caught by the new test
  restore the section-NAME compare -> caught by the new test AND by the moved
                                      licensing test

tools/test_tubuild.py -- the three moved tests removed (58 -> 55 functions),
and all sixteen `if not _toolchain(): return` guards converted to
`raise unittest.SkipTest(...)`. A bare return reports PASS having asserted
nothing; `pytest -rs` cannot see it and neither can a reviewer reading a green
log. This does NOT make the module wirable -- it still has no TestCase and a
third of it still compiles with mwccarm -- and the workflow comments that
implied otherwise are corrected.

.github/workflows/tool-tests.yml -- the new module added to the invocation
list (28 -> 29), plus the inventory entry, and every claim in the header that
this change falsified. tools.test_objisolate is deliberately still NOT wired:
it is a single @unittest.skipUnless(_compiler(), ...) class, so on a runner it
contributes 35 tests and 35 skips -- a green that asserts nothing, the exact
shape the header refuses a glob for.

Full CI invocation, run locally on a tree WITH the toolchain:

  Ran 519 tests in 61.934s
  OK (skipped=3)

check_python_names.py PASS. check_dead_references.py: no new dead references,
no broken markdown links.
